feat: Add reset password button
This button will use the form email if pressed,
if no email is present, it will ask for one.

When creating a user, the application will
redirect the user to login fragment with email
and password filled in, waiting for the validation mail.

This commit also lightly improve code on LoginFragment
- Remove useless constructor
- Avoid useless multiple calls on *EditText.getText()
- field content are not anymore attributes